+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - General usage documentation
 - Progress bar functionality for reading, writing, and copying operations
 - Support for Avro filetype
+- Regular expression matching functionality for `list_objects`
 
 ### Changed
 - Switched to standard NHDS CI pipeline